Conflicts:
	core/src/main/java/xdi2/core/features/nodetypes/XdiAbstractClass.java

+++ core/src/main/java/xdi2/core/features/nodetypes/XdiAbstractClass.java
@@ -126,7 +126,7 @@ public abstract class XdiAbstractClass<U extends XdiInstanceUnordered, O extends
 	 * Returns the number of XDI instances in this XDI class.
 	 */
 	@Override
-	public int getXdiInstancesUnorderedCount() {
+	public long getXdiInstancesUnorderedCount() {
 
 		return new IteratorCounter(this.getXdiInstancesUnordered()).count();
 	}
@@ -136,13 +136,13 @@ public abstract class XdiAbstractClass<U extends XdiInstanceUnordered, O extends
 	 * @return The XDI element.
 	 */
 	@Override
-	public O setXdiInstanceOrdered(int index) {
+	public O setXdiInstanceOrdered(long index) {
 
 		boolean attribute = this.attribute();
 
 		if (index < 0) index = this.getXdiInstancesOrderedCount();
 
-		XDI3SubSegment arcXri = XdiAbstractInstanceOrdered.createArcXri(Integer.toString(index), attribute);
+		XDI3SubSegment arcXri = XdiAbstractInstanceOrdered.createArcXri(Long.toString(index), attribute);
 
 		ContextNode contextNode = this.getContextNode().getContextNode(arcXri);
 		if (contextNode == null) contextNode = this.getContextNode().createContextNode(arcXri);
@@ -155,11 +155,11 @@ public abstract class XdiAbstractClass<U extends XdiInstanceUnordered, O extends
 	 * @return The XDI element.
 	 */
 	@Override
-	public O getXdiInstanceOrdered(int index) {
+	public O getXdiInstanceOrdered(long index) {
 
 		boolean attribute = this.attribute();
 
-		XDI3SubSegment arcXri = XdiAbstractInstanceOrdered.createArcXri(Integer.toString(index), attribute);
+		XDI3SubSegment arcXri = XdiAbstractInstanceOrdered.createArcXri(Long.toString(index), attribute);
 
 		ContextNode contextNode = this.getContextNode().getContextNode(arcXri);
 		if (contextNode == null) return null;
@@ -181,7 +181,7 @@ public abstract class XdiAbstractClass<U extends XdiInstanceUnordered, O extends
 	 * Returns the number of XDI elements in this XDI class.
 	 */
 	@Override
-	public int getXdiInstancesOrderedCount() {
+	public long getXdiInstancesOrderedCount() {
 
 		return new IteratorCounter(this.getXdiInstancesOrdered()).count();
 	}
